refactor(frontend): extract VaultTab (16/N)
Full vault UI: stats, default-resume banner, upload form, compare two
versions (TF-IDF), browse/search/sort all PDFs.

App.jsx 2186 → 1921 LOC (-265). Crosses the 2000-LOC line.
